add get_game_time() and get_last_tick(), transfer all time-funcs to seconds, fixes

+++ src/gui/screen/key.h
@@ -12,7 +12,7 @@ private:
 	static bool notReactOnEnter;
 	static std::vector<Key*> screenKeys;
 
-	long lastDown = 0;
+	double lastDown = 0;
 	int lastUpdate = -1;
 	bool prevIsDown = false;
